# Lammah-game-backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/questions": {
      "post": {
        "operationId": "questionsCreate",
        "summary": "Create a new question",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateQuestionDto"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "image": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  }
                ]
              },
              "examples": {
                "default": {
                  "summary": "Create approved free-game question",
                  "value": {
                    "category": "66b8f5f2c9d7a8b1e3f00222",
                    "question": "What planet is known as the Red Planet?",
                    "answer": "Mars",
                    "explanation": "Mars appears red because of iron oxide on its surface.",
                    "difficulty": "easy",
                    "points": 200,
                    "type": "text",
                    "status": "approved",
                    "source": "manual",
                    "isFreeGameQuestion": true
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateQuestionDto"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "image": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  }
                ]
              },
              "examples": {
                "default": {
                  "summary": "Create approved free-game question",
                  "value": {
                    "category": "66b8f5f2c9d7a8b1e3f00222",
                    "question": "What planet is known as the Red Planet?",
                    "answer": "Mars",
                    "explanation": "Mars appears red because of iron oxide on its surface.",
                    "difficulty": "easy",
                    "points": 200,
                    "type": "text",
                    "status": "approved",
                    "source": "manual",
                    "isFreeGameQuestion": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Question created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionMutationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": [
          "Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "questionsList",
        "summary": "Get all questions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Questions retrieved successfully. Answers are hidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Questions"
        ]
      }
    },
    "/questions/{id}": {
      "get": {
        "operationId": "questionsGetById",
        "summary": "Get a specific question by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Question MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00333",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Question retrieved successfully. Answer is hidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDetailResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Question not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Question with ID \"66b8f5f2c9d7a8b1e3f00333\" not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Questions"
        ]
      },
      "patch": {
        "operationId": "questionsUpdate",
        "summary": "Update a question",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Question MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00333",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionDto"
              },
              "examples": {
                "default": {
                  "summary": "Approve and mark as free-game question",
                  "value": {
                    "status": "approved",
                    "isFreeGameQuestion": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionMutationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Question not found"
          }
        },
        "tags": [
          "Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "questionsDelete",
        "summary": "Delete a question",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Question MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00333",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Question deleted successfully"
          },
          "404": {
            "description": "Question not found"
          }
        },
        "tags": [
          "Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/questions/accepted-answers/generate": {
      "post": {
        "operationId": "questionsGenerateAcceptedAnswers",
        "summary": "Generate reviewable accepted-answer aliases during authoring",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateAcceptedAnswersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedAnswerGenerationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/questions/accepted-answers/generate-ranked-list": {
      "post": {
        "operationId": "questionsGenerateRankedAcceptedAnswers",
        "summary": "Generate reviewable aliases for all ranked-list entries",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRankedAcceptedAnswersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankedAcceptedAnswerGenerationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/questions/check-duplicates": {
      "post": {
        "operationId": "questionsCheckDuplicates",
        "summary": "Check a manually-authored question for duplicates",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckQuestionDuplicatesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "exactMatch": false,
                    "highestSimilarity": 0.88,
                    "matches": []
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/questions": {
      "get": {
        "operationId": "adminQuestionsList",
        "summary": "Admin: get all questions including answers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Questions retrieved successfully with answers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/admin/questions/bulk-action": {
      "post": {
        "operationId": "questionsBulkAction",
        "summary": "Apply an admin review action to questions",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkQuestionActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkQuestionActionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/admin/questions/{id}/retry-primary-asset": {
      "post": {
        "operationId": "questionsRetryPrimaryAsset",
        "summary": "Retry primary asset resolution",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin Questions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Lammah Game API",
    "description": "Runtime HTTP contract for Lammah game administration, content, and gameplay.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "QuestionMutationResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/QuestionResponseDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "data",
          "message"
        ]
      },
      "QuestionListResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionResponseDto"
            }
          }
        },
        "required": [
          "statusCode",
          "data"
        ]
      },
      "QuestionDetailResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/QuestionResponseDto"
          }
        },
        "required": [
          "statusCode",
          "data"
        ]
      },
...
      "BulkQuestionActionResponseDto": {
        "type": "object",
        "properties": {
          "modifiedCount": {
            "type": "number"
          }
        },
        "required": [
          "modifiedCount"
        ]
      },
...
    }
  }
}